processesテーブルに関する処理をProcessControllerに移動
